# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 2Grils.PPT API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add a template to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{templateId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove an item from cart
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/voucher:
    post:
      operationId: applyVoucher
      tags: [cart]
      summary: Apply a voucher code to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyVoucherBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid voucher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Orders
...
  /custom-requests:
    get:
      operationId: listCustomRequests
      tags: [orders]
      summary: List user custom design requests
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomRequest"
    post:
      operationId: createCustomRequest
      tags: [orders]
      summary: Submit a custom design request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomRequestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomRequest"
  /custom-requests/{id}:
    get:
      operationId: getCustomRequest
      tags: [orders]
      summary: Get custom request details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomRequest"

  # Blog
  /blog:
    get:
      operationId: listBlogPosts
      tags: [templates]
      summary: List blog posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [templates]
      summary: Get blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  # Auth
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get currently authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Admin
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Cart:
      type: object
      required: [items, subtotal, discount, total, currency]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        currency:
          type: string
        appliedVoucher:
          type: ["string", "null"]
    CartItem:
      type: object
      required: [templateId, titleVi, titleEn, price, thumbnailUrl]
      properties:
        templateId:
          type: integer
        titleVi:
          type: string
        titleEn:
          type: string
        price:
          type: number
        thumbnailUrl:
          type: string
    AddToCartBody:
      type: object
      required: [templateId]
      properties:
        templateId:
          type: integer
    ApplyVoucherBody:
      type: object
      required: [code]
      properties:
        code:
          type: string
...
    CustomRequest:
      type: object
      required: [id, requestId, status, slideType, slideCount, deadline, createdAt, customerName, customerEmail]
      properties:
        id:
          type: integer
        requestId:
          type: string
        status:
          type: string
          enum: [pending, quoted, deposit-paid, in-progress, review, final-payment, delivered]
        slideType:
          type: string
        slideCount:
          type: integer
        deadline:
          type: string
        style:
          type: ["string", "null"]
        language:
          type: string
        budget:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        quotedPrice:
          type: ["number", "null"]
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: ["string", "null"]
        createdAt:
          type: string
    CreateCustomRequestBody:
      type: object
      required: [customerName, customerEmail, slideType, slideCount, deadline, language]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        company:
          type: string
        slideType:
          type: string
        targetAudience:
          type: string
        objective:
          type: string
        slideCount:
          type: integer
        style:
          type: string
        colorPalette:
          type: string
        aspectRatio:
          type: string
        language:
          type: string
        deadline:
          type: string
        budget:
          type: string
        notes:
          type: string
...
    BlogPost:
      type: object
      required: [id, slug, titleVi, titleEn, excerptVi, excerptEn, coverImageUrl, publishedAt, author]
      properties:
        id:
          type: integer
        slug:
          type: string
        titleVi:
          type: string
        titleEn:
          type: string
        excerptVi:
          type: string
        excerptEn:
          type: string
        contentVi:
          type: string
        contentEn:
          type: string
        coverImageUrl:
          type: string
        publishedAt:
          type: string
        author:
          type: string
        tags:
          type: array
          items:
            type: string
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, admin, designer]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    RegisterBody:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...